res fixed some VC7.0 warnings. res fixed crashes when making screenshots of render targets. res fixed text not showing up when drawing to a render target. res removed the clear-with-DrawBox() workarounf again. It caused   issues in other places, and the Clear() problem manifests itself   only in one place (PS labels), so it can be worked around there. res removed an entry from gldrivers.xml that became obsolete (and   non-functional) after the last texture manager changes (texture   class introduction etc.)
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@20755 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/video/render3d/software/sft3dcom.h
@@ -120,7 +120,7 @@ public:
 struct G3DPolygonDPFX
 {
   /// Current number of vertices.
-  int num;
+  size_t num;
   /// Vertices that form the polygon.
   csVector2 vertices[100];
   /// 1/z for every vertex.
@@ -157,7 +157,7 @@ struct G3DPolygonDPFX
 struct G3DPolygonDFP
 {
   /// Current number of vertices.
-  int num;
+  size_t num;
   /// Vertices that form the polygon.
   csVector2 vertices[100];
